tests/wpt/tests/css/css-animations/parsing/animation-direction-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-direction with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-direction">
<meta name="assert" content="animation-direction supports the full grammar '<single-animation-direction> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-direction", "normal");
test_valid_value("animation-direction", "reverse");
test_valid_value("animation-direction", "alternate");
test_valid_value("animation-direction", "alternate-reverse");
test_valid_value("animation-direction", "normal, reverse, alternate, alternate-reverse");
</script>
</body>
</html>
